Add: type files of components, containers
