+++ src/containers/CheckList/index.tsx
@@ -1,15 +1,9 @@
 import styled from "styled-components";
 import CheckBox from "../../components/CheckBox";
-import { CheckedListType } from "../Filters";
+import { CheckListProps } from "../../types/containers";
 
 const methodList = ["밀링", "선반"];
 const materialList = ["알루미늄", "탄소강", "구리", "합금강", "강철"];
-
-type CheckListProps = {
-  type: string;
-  handleCheck: (e: any) => void;
-  checkedList: CheckedListType;
-};
 
 function CheckList({ type, handleCheck, checkedList }: CheckListProps) {
   if (type === "method") {
